criado controller para abstrair o codigo e criado logica para verificar duplicidade de cadastro de dev
